+++ src/Services/Feed/Feed.Application/Article/Commands/VoteForArticle/ArticleRatingDto.cs
@@ -1,6 +1,6 @@
 ﻿namespace Feed.Application.Article.Commands.VoteForArticle {
     public class ArticleRatingDto {
-        public int Rating { get; init; }
+        public long Rating { get; init; }
         public short? Vote { get; init; }
     }
 }
